sign-conversion

+++ src/utils/nt_codec.h
@@ -32,7 +32,7 @@ inline auto nt_extract(char * seq, uint64_t pos) -> unsigned char
   static constexpr auto max_nt_per_uint64 = 32U;  // 32 nt fit in 64 bits
   static constexpr auto drop_remainder = 5U;  // (len+31) % 32 (drop remainder)
   static constexpr auto max_range = 3U;
-  const auto target_uint64 = pos >> drop_remainder;  // same as pos % 32
+  const auto target_uint64 = static_cast<long long int>(pos >> drop_remainder);  // same as pos % 32
   auto const * seq_uint64 = reinterpret_cast<uint64_t *>(seq);
   const auto compressed_chunk = *std::next(seq_uint64, target_uint64);
   // max_nt_per_uint64 - 1 = 31 = ...OOO1 1111 (mask all upper bits)
